26c1c34fd8144feaa3fb38fd96a0ef9244c7ca54 refactor: shed avoidable casts, follow-up to bitcoin#25611 (Konstantin Akimov)

Pull request description:

  ## Issue being fixed or feature implemented
  Small follow-up for bitcoin#25611 (merged via #6775)

  ## What was done?
  Shed avoidable casts in evo/mnhftx, llmq/debug

  ## How Has This Been Tested?
  See CI

  ## Breaking Changes
  N/A

  ## Checklist:
  - [x] I have performed a self-review of my own code
  - [ ] I have commented my code, particularly in hard-to-understand areas
  - [ ] I have added or updated relevant unit/integration/functional/e2e tests
  - [ ] I have made corresponding changes to the documentation
  - [x] I have assigned this pull request to a milestone

ACKs for top commit:
  UdjinM6:
    utACK 26c1c34fd8144feaa3fb38fd96a0ef9244c7ca54
  kwvg:
    utACK 26c1c34fd8144feaa3fb38fd96a0ef9244c7ca54

Tree-SHA512: 5b33360014ad449d95099bbb961e2cdcd00a9178cd4aac79b5e51367ab42d8a341dec085c753f4e2b7f9b5f8231ad24a4ea5b700d0a60f198f2702dec02d2591

+++ src/evo/mnhftx.h
@@ -50,7 +50,7 @@ public:
     [[nodiscard]] UniValue ToJson() const
     {
         UniValue obj(UniValue::VOBJ);
-        obj.pushKV("versionBit", (int)versionBit);
+        obj.pushKV("versionBit", versionBit);
         obj.pushKV("quorumHash", quorumHash.ToString());
         obj.pushKV("sig", sig.ToString());
         return obj;
